Add Promotional features for executive offices
Executive offices can have any number of "promotional features". Each
feature is made up of one or more "promotional feature items", which
consists of an image, some text, an optional title (with link) and up
to six additional links. Promotional feature items can be flagged as
being "double-width". These will be displayed with a larger image when
shown on frontend. The best way to understand this is probably by
checking out the visuals attached to the tracker story:
https://www.pivotaltracker.com/story/show/46261745